Load @word from the input parameters (URL) before trying to render the edit page
The reason this code needs to be here is that your form code says

`form_for @word ...`

which assumes that @word is defined at that time. However, since you didn't
ever write a `WordsController#edit` method, there was no code that was examining
the URL (/words/NNNN/edit) and loading up the NNNNth word and stuffing it into
the @word variable. As a result, the partial didn't know how to display the
form for @word because it was not yet defined.